add in non-working dropdown and basic dev panel
