+++ src/pages/Home.css
@@ -1,0 +1,144 @@
+.login { 
+	position: absolute;
+	width:300px;
+	height:300px;
+	cursor: default;
+}
+
+.student-side {
+	top: 30%;
+	left: 30%;
+	margin: -150px 0 0 -150px;
+}
+
+.parent-side {
+	top: 30%;
+	left: 80%;
+	margin: -150px 0 0 -150px;
+}
+
+.login h1 { color: #fff; text-shadow: 0 0 10px rgba(0,0,0,0.3); letter-spacing:1px; text-align:center; }
+
+input { 
+	width: 100%; 
+	margin-bottom: 10px; 
+	background: rgba(0,0,0,0.3);
+	border: none;
+	outline: none;
+	padding: 10px;
+	color: #fff;
+	text-shadow: 1px 1px 1px rgba(0,0,0,0.3);
+	border: 1px solid rgba(0,0,0,0.3);
+	border-radius: 4px;
+	box-shadow: inset 0 -5px 45px rgba(100,100,100,0.2), 0 1px 1px rgba(255,255,255,0.2);
+	-webkit-transition: box-shadow .5s ease;
+	-moz-transition: box-shadow .5s ease;
+	-o-transition: box-shadow .5s ease;
+	-ms-transition: box-shadow .5s ease;
+	transition: box-shadow .5s ease;
+}
+
+.form {
+	cursor: text;
+}
+
+input:focus { box-shadow: inset 0 -5px 45px rgba(100,100,100,0.4), 0 1px 1px rgba(255,255,255,0.2); }
+
+.names {
+	position: relative;
+	transition: border 0.2s ease-in;
+}
+
+.names::before {
+	content: "";
+	position: absolute;
+	top: 0;
+	left: 50%;
+	width: 0;
+	height: 2px;
+	background: rgba(255, 255, 255, 0.34);
+	transition: width 0.2s ease-in;
+	transform: translateX(-50%);
+}
+
+.names:hover::before {
+	width: 100%;
+}
+
+.name:hover {
+	cursor: pointer;
+}
+
+.name {
+	padding-top: 1vh;
+	padding-bottom: 1vh;
+	display: inline-flex;
+	align-items: center;
+	position: relative;
+	max-width: 0;
+	overflow: hidden;
+	animation: width-in 2s ease-out forwards;
+}
+
+.name span {
+	position: relative;
+}
+
+.name span::before {
+	content: "";
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	width: 0;
+	height: 3px;
+	background: white;
+	transition: width 0.3s ease-out;
+	transform: translateX(-50%);
+}
+
+.name span:hover {
+	color: rgba(255, 255, 255, 0.34);
+}
+
+.name span:hover::before {
+	width: 100%;
+}
+
+.students {
+	position:absolute;
+    top: 30%;
+    padding-left: 2vw;
+    font-size: 1.5rem;
+    width: 60%;
+    cursor: default;
+    animation: fade-in 1s ease-in forwards;
+}
+
+.mentors {
+	--left: 60%;
+	position:absolute;
+	top: 30%;
+	left: var(--left);
+	padding-left: 2vw;
+	font-size: 1.5rem;
+	width: calc(100% - var(--left));
+	cursor: default;
+}
+
+@keyframes fade-in {
+	from {
+		opacity: 0;
+	}
+	to {
+		opacity: 1;
+	}
+}
+
+@keyframes width-in {
+	from {
+		max-width: 0;
+	}
+	to {
+		max-width: 100%;
+	}
+}
